requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
aiosmtplib==2.0.2
alembic==1.13.2
amqp==5.2.0
annotated-types==0.7.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
assemblyai==0.32.0
astroid==3.2.4
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
Authlib==1.3.1
autopep8==2.3.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
billiard==4.2.0
black==24.4.2
bleach==6.1.0
blinker==1.8.2
Brotli==1.1.0
bs4==0.0.2
build==1.2.1
cachetools==5.4.0
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
cfgv==3.4.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
cryptography==43.0.0
cssselect==1.2.0
cssutils==2.11.1
dataclasses-json==0.6.7
decorator==4.4.2
deep-translator==1.11.4
deepgram-captions==1.2.0
deepgram-sdk==3.5.0
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.8
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.2
eventlet==0.36.1
exceptiongroup==1.2.2
Faker==26.0.0
fastapi==0.111.1
fastapi-cli==0.0.4
fastapi-mail==1.4.1
fastapi-users==13.0.0
ffmpeg-python==0.2.0
ffprobe==0.5
filelock==3.15.4
flake8==7.1.0
flatbuffers==24.3.25
flower==2.0.1
fpdf==1.7.2
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-cloud-speech==2.27.0
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.4
grpcio-status==1.65.4
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.8.1
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.6
humanfriendly==10.0
humanize==4.10.0
hyperframe==5.2.0
identify==2.6.0
idna==2.10
imageio==2.35.0
imageio-ffmpeg==0.5.1
importlib_resources==6.4.4
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.4.0
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.32
langchain-openai==0.1.21
langchain-text-splitters==0.2.2
langsmith==0.1.99
lazy_loader==0.4
librosa==0.10.2.post1
limits==3.13.0
llvmlite==0.43.0
logger==1.4
lxml==5.2.2
makefun==1.15.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mccabe==0.7.0
mdurl==0.1.2
minio==7.2.8
more-itertools==10.3.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mutagen==1.47.0
mypy-extensions==1.0.0
networkx==3.3
nodeenv==1.9.1
numba==0.60.0
numpy==1.26.4
onnxruntime==1.18.1
openai==1.40.3
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
orjson==3.10.7
packaging==24.1
pandas==2.2.2
passlib==1.7.4
pathspec==0.12.1
pillow==10.4.0
pip-tools==7.4.1
pipdeptree==2.23.1
platformdirs==4.2.2
pluggy==1.5.0
pooch==1.8.2
pre-commit==3.7.1
premailer==3.10.0
proglog==0.1.10
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.3
psycopg2-binary==2.9.9
pwdlib==0.2.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclipper==1.3.0.post5
pycodestyle==2.12.0
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
pydub==0.25.1
pyee==11.1.0
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.8.0
pylint==3.2.6
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
PyMySQL==1.1.1
pyparsing==3.1.2
pypdf==4.3.1
pyproject_hooks==1.1.0
pyreadline3==3.4.1
pytesseract==0.3.10
pytest==8.2.2
pytest-asyncio==0.23.8
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytubefix==6.10.2
pytz==2024.1
PyYAML==6.0.1
rapidocr-onnxruntime==1.3.24
redis==5.0.8
regex==2024.7.24
reportlab==4.2.2
requests==2.32.3
rfc3986==1.5.0
rich==13.7.1
rsa==4.9
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.0
shapely==2.0.5
shellingham==1.5.4
six==1.16.0
slowapi==0.1.9
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.6
soxr==0.4.0
SpeechRecognition==3.10.4
SQLAlchemy==2.0.31
sse-starlette==2.1.3
starlette==0.37.2
stripe==10.7.0
sympy==1.13.2
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.13.0
tornado==6.4.1
tqdm==4.66.5
transformers==4.44.2
twilio==9.2.3
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uuid7==0.1.0
uvicorn==0.30.3
uvloop==0.19.0
vine==5.1.0
virtualenv==20.26.3
watchdog==4.0.2
watchfiles==0.22.0
Wave==0.0.2
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
yt-dlp==2024.8.6
